requirements.txt

aiohappyeyeballs==2.6.1; python_version >= '3.9'
aiohttp==3.11.14; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
altair==5.5.0; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
async-timeout==4.0.3; python_version >= '3.7'
attrs==25.3.0; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
cachetools==5.5.2; python_version >= '3.7'
certifi==2025.1.31; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
cryptography==44.0.2; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
dataclasses-json==0.6.7; python_version >= '3.7' and python_version < '4.0'
distro==1.9.0; python_version >= '3.6'
exceptiongroup==1.2.2; python_version >= '3.7'
faiss-cpu==1.10.0; python_version >= '3.9'
fastapi==0.115.11; python_version >= '3.8'
frozenlist==1.5.0; python_version >= '3.8'
gitdb==4.0.12; python_version >= '3.7'
gitpython==3.1.44; python_version >= '3.7'
greenlet==3.1.1; python_version >= '3.7'
groq==0.20.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
httpx-sse==0.4.0; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jinja2==3.1.6; python_version >= '3.7'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==3.0.0; python_version >= '3.7'
jsonschema==4.23.0; python_version >= '3.8'
jsonschema-specifications==2024.10.1; python_version >= '3.9'
langchain==0.3.21; python_version >= '3.9' and python_version < '4.0'
langchain-community==0.3.20; python_version >= '3.9' and python_version < '4.0'
langchain-core==0.3.47; python_version >= '3.9' and python_version < '4.0'
langchain-groq==0.3.1; python_version >= '3.9' and python_version < '4.0'
langchain-ollama==0.3.0; python_version >= '3.9' and python_version < '4.0'
langchain-text-splitters==0.3.7; python_version >= '3.9' and python_version < '4.0'
langgraph==0.3.18; python_full_version >= '3.9.0' and python_version < '4.0'
langgraph-checkpoint==2.0.21; python_full_version >= '3.9.0' and python_full_version < '4.0.0'
langgraph-prebuilt==0.1.4; python_full_version >= '3.9.0' and python_full_version < '4.0.0'
langgraph-sdk==0.1.58; python_full_version >= '3.9.0' and python_full_version < '4.0.0'
langsmith==0.3.18; python_version >= '3.9' and python_version < '4.0'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.26.1; python_version >= '3.9'
msgpack==1.1.0; python_version >= '3.8'
multidict==6.2.0; python_version >= '3.9'
mypy-extensions==1.0.0; python_version >= '3.5'
narwhals==1.31.0; python_version >= '3.8'
numpy==2.0.2; python_version >= '3.9'
ollama==0.4.7; python_version >= '3.8' and python_version < '4.0'
orjson==3.10.15; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pandas==2.2.3; python_version >= '3.9'
pdfminer.six==20231228; python_version >= '3.6'
pdfplumber==0.11.5; python_version >= '3.8'
pillow==11.1.0; python_version >= '3.9'
propcache==0.3.0; python_version >= '3.9'
protobuf==5.29.4; python_version >= '3.8'
pyarrow==19.0.1; python_version >= '3.9'
pycparser==2.22; python_version >= '3.8'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pydantic-settings==2.8.1; python_version >= '3.8'
pydeck==0.9.1; python_version >= '3.8'
pypdfium2==4.30.1; python_version >= '3.6'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2025.1
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
requests==2.32.3; python_version >= '3.8'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rpds-py==0.23.1; python_version >= '3.9'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
smmap==5.0.2; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.39; python_version >= '3.7'
starlette==0.46.1; python_version >= '3.9'
streamlit==1.43.2; python_version >= '3.9' and python_full_version != '3.9.7'
tenacity==9.0.0; python_version >= '3.8'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
tornado==6.4.2; python_version >= '3.8'
typing-extensions==4.12.2; python_version >= '3.8'
typing-inspect==0.9.0
tzdata==2025.1; python_version >= '2'
urllib3==2.3.0; python_version >= '3.9'
uvicorn==0.34.0; python_version >= '3.9'
watchdog==6.0.0; python_version >= '3.9'
yarl==1.18.3; python_version >= '3.9'
zstandard==0.23.0; python_version >= '3.8'
